requirements.txt
groq==0.13.1
httpx==0.27.2
Pillow==11.0.0
python-dotenv==1.0.1
requests==2.32.3
tenacity==9.0.0
google-api-python-client==2.151.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
